# docs/api/openapi-full.yaml
openapi: 3.1.0
info:
  title: DataSafeS3 REST Admin API
  version: 1.1.0
  summary: Complete JSON REST API served at /api/v1 (excluding S3 XML)
  description: |
    **Full hand-written OpenAPI 3.1** for all Admin JSON routes registered in internal/api/server.go.

    ## Authentication
    | Scheme | Use |
    |--------|-----|
    | **BearerAPIToken** | Integration tokens ds_* (recommended for automation) |
    | **BearerJWT** | JWT from POST /admin/login or OIDC callback (console sessions) |

    Public endpoints: GET /health, POST /admin/login, OIDC bootstrap, GET /public/share/*.

    ## Not included
    - GET /metrics (Prometheus text)
    - S3 XML API on paths outside /api/v1 (SigV4, port 9000)

    ## Errors
    Standard envelope: JSON object with error field (HTTP 4xx/5xx).

    Interactive docs for **Community subset**: GET /api/v1/docs (Swagger UI).

    Canonical file: [openapi-full.yaml](../../api/openapi-full.yaml)
  contact:
    name: DataSafeS3 / Ilya Trachuk
    url: https://github.com/DirektorBani/DataSafeS3
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Via Caddy (console origin)
  - url: http://localhost:9000/api/v1
    description: storage-server direct
security:
  - BearerAPIToken: []
  - BearerJWT: []
components:
  securitySchemes:
    BearerAPIToken:
      type: http
      scheme: bearer
      bearerFormat: ds_
      description: |
        Integration API token with prefix ds_. Create in console: Access → API tokens.
        Inherits creator's role and bucket access.

    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT from POST /admin/login, MFA completion, or OIDC callback.
  parameters:
    BucketName:
      name: bucket
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        mfa_required: { type: boolean }
        mfa_token: { type: string }
        mfa_setup_required: { type: boolean }
        scheduled_delete_at: { type: string, format: date-time }
        trashed: { type: boolean }
        trash_id: { type: string }
...
    ShareLink:
      type: object
      properties:
        id: { type: string }
        bucket: { type: string }
        key: { type: string }
        token: { type: string }
        max_downloads: { type: integer }
        download_count: { type: integer }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
...
    CreateShareRequest:
      type: object
      required: [key]
      properties:
        key: { type: string, description: Object key to share }
        expires_in_sec: { type: integer, minimum: 0, description: "TTL; 0 = no expiry" }
        max_downloads: { type: integer, minimum: 0, default: 0 }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
paths:
...
  /users/{id}:
    put:
      operationId: updateUser
      summary: Update user
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200': { description: Updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      operationId: deleteUser
      summary: Delete user
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /buckets/{bucket}/shares:
    get:
      operationId: listShares
      summary: List share links
      tags: [Shares]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - $ref: '#/components/parameters/BucketName'
          - name: key
            in: query
            schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  shares:
                    type: array
                    items: { $ref: '#/components/schemas/ShareLink' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    post:
      operationId: createShare
      summary: Create share link
      tags: [Shares]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - $ref: '#/components/parameters/BucketName'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateShareRequest' }
      responses:
        '201':
          description: Share created (expires_in_sec in request body)
          content:
            application/json:
              schema:
                type: object
                properties:
                  share: { $ref: '#/components/schemas/ShareLink' }
                  url: { type: string, format: uri }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /shares/{id}:
    delete:
      operationId: revokeShare
      summary: Revoke share link
      tags: [Shares]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /public/share/{token}:
    get:
      operationId: publicShareInfo
      summary: Public share metadata
      tags: [Shares]
      security: []
      parameters:
          - name: token
            in: path
            required: true
            schema: { type: string }
      responses:
        '200': { description: Share info (no auth) }
...
  /tenants/{id}/members:
    get:
      operationId: listTenantMembers
      summary: List tenant members
      tags: [Tenants]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200': { description: Members }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    post:
      operationId: addTenantMember
      summary: Add tenant member
      tags: [Tenants]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '201': { description: Added }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...